feat: separate app.ts file
